# microservice/contracts/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Content/Admin Service API
  version: 0.1.0
  summary: Catalog management endpoints
  description: 管理者向けに書籍データの登録・更新・公開制御を提供する。
servers:
  - url: https://api.bookstore.local/admin
paths:
  /books:
    post:
      summary: 書籍を登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertBookRequest'
      responses:
        '201':
          description: 登録された書籍
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
  /books/{bookId}:
    get:
      summary: 書籍を取得
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '200':
          description: 書籍詳細
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: 書籍を更新
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertBookRequest'
      responses:
        '200':
          description: 更新後の書籍
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 書籍を削除
      parameters:
        - $ref: '#/components/parameters/BookId'
      responses:
        '204':
          description: 削除完了
  /books/{bookId}/publish:
    post:
      summary: 書籍の公開状態を変更
      parameters:
        - $ref: '#/components/parameters/BookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/PublicationStatus'
      responses:
        '200':
          description: 更新後の公開状態
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/BookDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: string
  schemas:
    PublicationStatus:
      type: string
      enum:
        - draft
        - published
        - archived
    BookDetail:
      type: object
      required:
        - id
        - title
        - author
        - price
        - status
        - description
        - fullDescription
        - publicationStatus
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        price:
          type: number
        status:
          type: string
        publicationStatus:
          $ref: '#/components/schemas/PublicationStatus'
        description:
          type: string
        fullDescription:
          type: string
        tags:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
        imageUrl:
          type: string
          format: uri
          nullable: true
    UpsertBookRequest:
      type: object
      required:
        - title
        - author
        - price
        - status
        - description
        - fullDescription
      properties:
        title:
          type: string
        author:
          type: string
        price:
          type: number
          minimum: 0
        status:
          type: string
          enum:
            - available
            - coming-soon
            - sold-out
        description:
          type: string
        fullDescription:
          type: string
        tags:
          type: array
          items:
            type: string
        publicationYear:
          type: integer
        imageUrl:
          type: string
          format: uri
          nullable: true
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: データ未検出
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'